moved all models into separate files in the models folder and moved the review interface to definitions
